requirements.txt
xrdtools>=0.1.0
scipy>=1.11.0
numpy>=1.24.0
requests>=2.31.0
gradio>=4.0.0
pyyaml>=6.0
